Move `CommentLoaderPresentationAdapter` to its own file
